refactor: move status and event choices to separate package
